Add files via upload
As a system engineer your expertise is asked to create a firewall ruleset for a hosting server.

The server is provided with the following services: Apache, ProFTPd and bind9. Please, do not allow zonetransfers. Also protect the server against ping flooding. The server is not allowed to make outgoing connections, except for the installation of security updates.